Address PR comments. Remove duplicated files
